esta bueno antes de paola

+++ modifyUsers.php
@@ -3,13 +3,17 @@ session_start();
 if (!isset($_SESSION['user_id'])) {
   header("Location: index.php");
   exit();
-}
-
-include("conexion.php");
+} else if ($_SESSION['is_admin'] == 1 || $_SESSION['is_admin'] == 2) {
+  include("conexion.php");
 $con = connection();
 
-$sql = "SELECT * FROM users";
-$query = mysqli_query($con, $sql);
+if ($_SESSION['is_admin'] == 1) {
+  $sql = "select users.*, admins.is_admin from users left join admins on users.id = admins.user_id";
+  $query = mysqli_query($con, $sql);
+} else {
+  $sql = "select users.*, admins.is_admin from users left join admins on users.id = admins.user_id where admins.is_admin = 0";
+  $query = mysqli_query($con, $sql);
+}
 ?>
 
 <!DOCTYPE html>
@@ -18,14 +22,22 @@ $query = mysqli_query($con, $sql);
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="css/mainPage.css">
-  <title>Vals Coffee | Editar</title>
+  <title>Vals Coffee | Actualizar</title>
 </head>
 <body>
   <div id="dashboard">
     <a href="showUsers.php" class="tab">Mostrar Usuario</a>
-    <a class="tab active">Modificar Usuario</a>
-    <a href="deleteUsers.php" class="tab">Eliminar Usuario</a>
-    <a href="printUsers.php" target="_blank" class="tab">Imprimir Usuarios</a>
+    <?php if ($_SESSION['is_admin'] == 1): ?>
+      <a href='createUser.php' class='tab'>Crear Usuario</a>
+      <a href='modifyUsers.php' class='tab active'>Modificar Usuario</a>
+      <a href='deleteUsers.php' class='tab'>Eliminar Usuario</a>
+      <a href="printUsers.php" target="_blank" class="tab">Imprimir Usuarios</a>
+    <?php elseif ($_SESSION['is_admin'] == 2): ?>
+      <a href='createUser.php' class='tab'>Crear Usuario</a>
+      <a href='modifyUsers.php' class='tab'>Actualizar Usuario</a>
+      <a href="printUsers.php" target="_blank" class="tab">Imprimir Usuarios</a>
+    <?php endif; ?>
+    <a href="myProfile.php" class="tab">Mi Perfil</a>
     <a href="logOut.php" class="tab">Cerrar Sesión</a>
   </div>
   <div class="content">
@@ -61,3 +73,8 @@ $query = mysqli_query($con, $sql);
   </div>
 </body>
 </html>
+<?php } else {
+  header("Location: index.php");
+  exit();
+}
+?>
